feat: implement Services for each exercise
